Move trinity config test file
